Bump version odlparent->6.0.0,mdsal->5.0.3
This adopts new version of MRI projects and adjusts for the changes
therein. Most notably DPs/DPNs are identified by Uint64, not
BigInteger.

Some drive-by cleanups are performed, mostly to eliminate valid
Eclipse complaints.

+++ commons/testutils/src/main/java/org/opendaylight/genius/testutils/TestInterfaceManager.java
@@ -9,7 +9,6 @@ package org.opendaylight.genius.testutils;
 
 import static org.opendaylight.yangtools.testutils.mockito.MoreAnswers.realOrException;
 
-import java.math.BigInteger;
 import java.util.Map;
 import java.util.Optional;
 import java.util.concurrent.ConcurrentHashMap;
@@ -24,6 +23,7 @@ import org.opendaylight.genius.testutils.interfacemanager.InterfaceHelper;
 import org.opendaylight.genius.testutils.interfacemanager.InterfaceStateHelper;
 import org.opendaylight.genius.testutils.interfacemanager.TunnelInterfaceDetails;
 import org.opendaylight.yang.gen.v1.urn.ietf.params.xml.ns.yang.ietf.interfaces.rev140508.interfaces.Interface;
+import org.opendaylight.yangtools.yang.common.Uint64;
 
 /**
  * IInterfaceManager implementation for tests.
@@ -147,12 +147,12 @@ public abstract class TestInterfaceManager implements IInterfaceManager {
     }
 
     @Override
-    public BigInteger getDpnForInterface(String interfaceName) {
+    public Uint64 getDpnForInterface(String interfaceName) {
         return interfaceInfos.get(interfaceName).getDpId();
     }
 
     @Override
-    public BigInteger getDpnForInterface(Interface intrface) {
+    public Uint64 getDpnForInterface(Interface intrface) {
         return interfaceInfos.get(intrface.getName()).getDpId();
     }
